feat(parser combinators): behave like classic combinators
for now, just use lists as is done on master, to preserve the exact
semantics. We still do provide the new parser combinators.
As we go, we might opt into other underlying implementations for
`repeating`, `req`, and `opt` (e.g. an array of pairs, and a bitvector
to indicate which pairs have been consumed).